+++ ulid_uint128.hh
@@ -142,7 +142,7 @@ inline void EncodeEntropyRand(ULID& ulid) {
 	ulid |= e;
 }
 
-static std::uniform_int_distribution<uint8_t> Distribution_0_255(0, 255);
+static std::uniform_int_distribution<uint32_t> Distribution_0_255(0, 255);
 
 /**
  * EncodeEntropyMt19937 will encode a ulid using std::mt19937
